+++ changelog.md
@@ -6,6 +6,8 @@
 
 - Updated Node.js support to `^12.20.0 || ^14.13.1 || >= 16.0.0`.
 - Updated dev dependencies, some of which require newer Node.js versions than previously supported.
+- Updated the [`graphql`](https://npm.im/graphql) peer dependency to `^16.0.0`.
+- Updated the `errorHandler` Koa middleware to avoid the `formatError` function deprecated in [`graphql`](https://npm.im/graphql) v16, using the new `GraphQLError.toJSON` method.
 
 ### Patch